+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN3FZ0CKRYMM6G274CRVVF2Y"
+        "01KN46CKVA5DWF4JP2PNMK6S0P"
     )
     r"""The unique identifier of the data source"""