+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN3FZ0CPKM22K8DFEVQBCZJ5"
+        "01KN46CKVDCA0E5ZR3FVSADX49"
     )
     r"""The unique identifier of the data source"""