+++ packages/orq-rc/RELEASES.md
@@ -6769,3 +6769,13 @@ Based on:
 - [python v4.7.0-rc.36] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.36] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.36 - packages/orq-rc
+
+## 2026-04-01 09:38:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.38] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.38] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.38 - packages/orq-rc